+++ src/aks-preview/HISTORY.rst
@@ -13,6 +13,10 @@ Pending
 ++++++
 * Vendor new SDK and bump API version to 2023-05-02-preview.
 
+0.5.144
++++++++
+* Fix setup network profile with network observability due to incorrect property
+
 0.5.143
 +++++++
 * Add `--enable-network-observability` flag to `az aks create` and `az aks update`.